Move yac_vote_storage to own folder
