+++ irohad/consensus/yac/storage/yac_vote_storage.hpp
