+++ irohad/consensus/yac/yac.hpp
@@ -28,7 +28,7 @@
 #include "consensus/yac/yac_network_interface.hpp"
 #include "consensus/yac/yac_crypto_provider.hpp"
 #include "consensus/yac/timer.hpp"
-#include "consensus/yac/yac_vote_storage.hpp"
+#include "consensus/yac/storage/yac_vote_storage.hpp"
 
 namespace iroha {
   namespace consensus {
